titan-poss-poss-web-f78a2c9ef984/titan-poss-poss-web-f78a2c9ef984/libs/eposs/gst-mapping/data-access-gst-mapping/src/lib/+state/gst-mapping.selector.ts
import { createSelector } from '@ngrx/store';
import { selectGSTMappingState } from './gst-mapping.reducer';

const selectError = createSelector(selectGSTMappingState, state => state.error);
const selectGSTMappingList = createSelector(
  selectGSTMappingState,
  state => state.gstMappingList
);
const selectTotalElements = createSelector(
  selectGSTMappingState,
  state => state.totalElements
);

const selectIsLoading = createSelector(
  selectGSTMappingState,
  state => state.isLoading
);

const selectTxnTypes = createSelector(
  selectGSTMappingState,
  state => state.txnTypes
);

const selectReloadStatus = createSelector(
  selectGSTMappingState,
  state => state.reloadStatus
);

const selectTaxes = createSelector(selectGSTMappingState, state => state.taxes);

export const GSTMappingSelectors = {
  selectError,
  selectGSTMappingList,
  selectIsLoading,
  selectTotalElements,
  selectTxnTypes,
  selectTaxes,
  selectReloadStatus,
};
